+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - MinIO storage creation now done in deltafi-core instead of helm
 - Flow assignment rules now has a name that is independent of the rule ID
 - Removed "egressFlow" parameter for EgressActions.
+- Changed "minio.expiration-days" property to be editable and refreshable
 
 ### Deprecated
 - Quarkus is no longer in use or supported in the DeltaFi monolith